Trying to create variables that will appear for the user
So they can be used to connect sound generators to FX, etc.
This may not be the right paradigm to use - we may have
to make the FX use input_statements like the Run block -
problem there is we have to find out the variable name by looking at the code.
The other option is an input_value field - like with variables - set x to:
Then the code generator may have to return something else, but don't know how...